+++ noxfile.py
@@ -51,6 +51,7 @@ SYSTEM = platform.system().lower()
 REQUIREMENTS_FORMAT = ['yapf==0.27']
 REQUIREMENTS_LINT = ['yapf==0.27', 'pyflakes==2.4.0', 'pytest==6.2.5', 'liccheck==0.6.2']
 REQUIREMENTS_TEST = [
+    "pexpect==4.8.0",
     "pytest==6.2.5",
     "pytest-cov==3.0.0",
     'pytest-xdist==2.5.0',
@@ -61,7 +62,7 @@ REQUIREMENTS_BUILD = ['setuptools>=20.2']
 REQUIREMENTS_BUNDLE = [
     'pyinstaller==5.6.2',
     "patchelf-wrapper==1.2.0;platform_system=='Linux'",
-    "staticx==0.13.5;platform_system=='Linux'",
+    "staticx==0.13.8;platform_system=='Linux'",
 ]
 
 WINDOWS_TIMESTAMP_SERVER = 'http://timestamp.digicert.com'