+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Add `install-autocomplete` command for installing shell autocompletion
+
 ### Infrastructure
 * GitHub CD builds and uploads an official B2 CLI image to docker hub
 * Disable changelog verification for dependabot PRs