+++ CHANGELOG.md
@@ -6,8 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.8.0] - 2023-03-23
+
 ### Added
 * Add `install-autocomplete` command for installing shell autocompletion
+
+### Fixed
+* Hitting the download endpoint twice in some cases
 
 ### Infrastructure
 * GitHub CD builds and uploads an official B2 CLI image to docker hub
@@ -407,8 +412,9 @@ This version is pinned strictly to `b2-sdk-python==1.16.0` for the same reason.
 * Fix an off-by-one issue when downloading a range of a file (affects library, but not CLI).
 * Better handling of some errors from the B2 service.
 
-[Unreleased]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.7.1...HEAD
-[3.7.1]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.7.1...v3.7.1
+[Unreleased]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.8.0...HEAD
+[3.8.0]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.7.1...v3.8.0
+[3.7.1]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.7.0...v3.7.1
 [3.7.0]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.6.0...v3.7.0
 [3.6.0]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.5.0...v3.6.0
 [3.5.0]: https://github.com/Backblaze/B2_Command_Line_Tool/compare/v3.4.0...v3.5.0